Examine information about dolt tables and diffs

Examine information about dolt tables and diffs

dolt_diffs(table, to, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_table_history(table, conn = dolt(), collect = NULL, show_sql = NULL)

Arguments

table
character the name of a table in the database

to
commit to compare to

conn
the database connection

collect
whether to collect the result into R or return a dbplyr::tbl_lazy() to be further processed before collecting. Defaults to TRUE, can be set with the environment variable DOLT_COLLECT.

show_sql
Whether to print the SQL statements used internally to fetch the data. Useful for learning how Dolt works internally. Defaults to FALSE, can be set with the environment variable DOLT_VERBOSE.