Open a Dolt connection pane in RStudio

This function launches the RStudio "Connection" pane to interactively explore the database. The pane will show the database versioning state, tables stored in the database, and dolt system tables showing history.

dolt_pane(conn = dolt())

update_dolt_pane(conn = dolt())

close_dolt_pane(conn = dolt())

Arguments

conn
a dolt connection. If a path is provided instead, a connection will be created to the path using dolt().

Value

The connection object (invisibly)

Details

When running dolt interactively, the connection pane will automatically update in response to most queries that modify the database state. You can stop this behavior by setting the DOLT_WATCH environment variable to 0 or false. See dolt_vars for more configuration variables