Connect to a dolt database

dolt_remote() is a DBI Driver to connect to a remote dolt server via a port. It, DoltDriver, and DoltConnection class are wrappers around the around classes and methods from the RMariaDB package.

Most parameters can be specified with package configuration environment variables.

dolt_remote()

# S4 method for DoltDriver
dbUnloadDriver(drv, ...)

# S4 method for DoltDriver
show(object)

# S4 method for DoltDriver
dbConnect(
  drv = dolt_remote(),
  dbname = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  port = Sys.getenv("DOLT_PORT", 3306L),
  autocommit = TRUE,
  ...
)

Arguments

drv

an object of class DoltDriver, created by dolt_remote().

...

other arguments passed to RMariaDB::MariaDB

object

a connection object

dbname

The database name

username

The username. Defaults to "root"

password

The login password. Defaults to empty.

host

The IP of the host. Defaults to the local machine, 127.0.0.1

port

The TCP port for connections. Defaults to 3306.

autocommit

Whether to autocommit changes in the SQL sense. That is, to flush pending changes to disk and update the working set.

Details

Most methods fall back to those for RMariaDB.

See also