Work with dolt repository remotes

Work with dolt repository remotes

dolt_push(
  remote = NULL,
  ref = NULL,
  set_upstream = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_pull(
  remote = NULL,
  squash = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_fetch(
  remote = NULL,
  ref = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

Arguments

remote

the name of the remote. "origin" is used by default

ref

the branch reference

set_upstream

whether to set the remote branch reference to track

force

whether to overwrite any conflicting history the current branch

conn

The dolt database to query

collect

Whether to return the full results (TRUE) or a an uncollected dplyr::tbl() to be further processed by dplyr pipelines before collecting. Can be set globally with the environment variable DOLT_COLLECT.

show_sql

Whether to print to the console the SQL statement executed. Useful in learning dolt SQL syntax. Defaults to false. Can be set globally with the environment variable DOLT_SHOW_SQL.

squash

whether to merge changes to the working set without updating the commit history

See also