Work with dolt repository remotes

Work with dolt repository remotes

dolt_push(
  remote = NULL,
  remote_branch = NULL,
  ref = NULL,
  set_upstream = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_pull(
  remote = NULL,
  squash = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_fetch(
  remote = NULL,
  ref = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_clone(
  remote_url,
  remote = "origin",
  new_dir = basename(remote_url),
  branch = NULL
)

Arguments

remote

the name of the remote. "origin" is used by default

remote_branch

the name of the remote branch to use with set_upstream. Current local branch is used by default

ref

the branch reference

set_upstream

whether to set the remote branch reference to track

force

whether to overwrite any conflicting history the current branch

conn

the database connection

collect

whether to collect the result into R or return a dbplyr::tbl_lazy() to be further processed before collecting. Defaults to TRUE, can be set with the environment variable DOLT_COLLECT.

show_sql

Whether to print the SQL statements used internally to fetch the data. Useful for learning how Dolt works internally. Defaults to FALSE, can be set with the environment variable DOLT_VERBOSE.

squash

whether to merge changes to the working set without updating the commit history

remote_url

the remote URL to clone

new_dir

the directory to clone into

branch

the branch to clone. If NULL, clones all branches