Connect to a local dolt database directory

dolt_local() creates a DoltLocalDriver, which can generate a DoltLocalConnection. Unlike dolt_remote() and DoltDriver, this version takes a directory name for an on-disk dolt database, starts and manages a dolt SQL server in the background serving that directory, connects to it and returns the connection. Parameters govern both the server and connection

Multi-user or other, more complicated networking set-ups should use dolt_server() and dolt_remote() directly.

DBI::dbConnect(doltr::dolt_local(), ...)

# S4 method for DoltLocalDriver
dbConnect(
  drv,
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  port = Sys.getenv("DOLT_PORT", 3306L),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  find_port = TRUE,
  server_args = list(),
  ...
)

Arguments

drv

an object of class DoltLocalDriver, created by dolt_local().

dir

The dolt directory to serve and connect to

username

The username. Defaults to "root"

password

The login password. Defaults to empty.

port

The TCP port for connections. Defaults to 3306.

host

The IP of the host. Defaults to the local machine, 127.0.0.1

server_args

a list of additional arguments to pass to dolt_server()

...

additional arguments to pass to RMariaDB

See also